Fixed problem with int/float conflict, so now instead of the entire app crashing it just kicks you out of settings.

+++ app/src/main/java/edu/rpi/shuttletracker/SettingsActivity.kt
@@ -115,7 +115,7 @@ public class SettingsActivity : AppCompatActivity() {
         discreteSlider.setOnSeekBarChangeListener(object : SeekBar.OnSeekBarChangeListener {
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
                 val editor = sharedPreferences.edit()
-                editor.putInt("min_stop_dist", progress)
+                editor.putFloat("min_stop_dist", progress.toFloat())
             }
 
             override fun onStartTrackingTouch(seekBar: SeekBar?) {
